added readthedocs link to readme, fixed link on deployment-instructions, and moved plot_bbox to visualization
