+++ .deployment-instructions.md
@@ -8,7 +8,7 @@
 
 ### Uploading a package to PyPI
 
-The link in this tutorial will explain the steps to upload a package to pypi: <https://dzone.com/articles/executable-package-pip-install.>
+The link in this tutorial will explain the steps to upload a package to pypi: <https://dzone.com/articles/executable-package-pip-install>
 
 #### Get started with Travis CI
